Rename  package to  with generic funcs
